project2/requirements.txt
pyarrow
pandas
psycopg2-binary
